requirements.txt
certifi==2026.1.4
chardet==5.2.0
#coreapi==2.3.3
#coreschema==0.0.4
Django==6.0.1
django-bootstrap3==26.1
#django-bower==5.2.0
django-cors-headers==4.9.0
django-countries==5.0
django-filter==25.2
drf-yasg==1.21.14
django-split-settings==1.3.2
djangorestframework==3.16.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.3
Markdown==3.10
MarkupSafe==3.0.3
nodeenv==1.10.0
openapi-codec==1.3.2
psycopg2==2.9.11
pytz==2025.2
requests==2.26.0
rest-framework-ember==1.3.2
simplejson==3.13.2
six==1.17.0
sqlparse==0.5.5
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.26.6
setuptools==80.9.0
legacy-cgi==2.6.4